requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
earthengine-api==0.1.323
numpy==1.21.2
Pillow==8.3.2
